+++ changelog.md
@@ -1,2 +1,2 @@
 ### Item Collectors 1.1.11
-- Initial release of Item Collectors for Minecraft 1.21.9 & 1.21.10
+- Initial release of Item Collectors for Minecraft 1.21.11